refactor(ui): extract project memory skills panels
