+++ ui/src/features/projects/ProjectsView.tsx
@@ -81,6 +81,8 @@ import { ProjectAssistantPanel } from "./ProjectAssistantPanel";
 import { EditAssignmentModal, NewAssignmentModal } from "./ProjectAssignmentModals";
 import { CreateProjectWorktreeModal } from "./CreateProjectWorktreeModal";
 import { ProjectHandoffModal } from "./ProjectHandoffModal";
+import { ProjectMemoryModal, ProjectMemoryPanel, type MemoryForm } from "./ProjectMemoryPanel";
+import { ProjectSkillsPanel } from "./ProjectSkillsPanel";
 import { ProfilesModal } from "./ProfilesModal";
 import { EditWorkItemModal, NewWorkItemModal } from "./ProjectWorkItemModals";
 import { RolesModal } from "./RolesModal";
@@ -91,7 +93,6 @@ import type {
   ProjectActivityItemRecord,
   ProjectMemoryCandidateRecord,
   ProjectCollaborationArtifactRecord,
-  ProjectContextSourceRecord,
   CreateProjectWorktreeRootPayload,
   ProjectHandoffRecord,
   ProjectMemoryRecord,
@@ -187,40 +188,6 @@ type AssignmentLaunchRepairActions = {
   onOpenConnections?: () => void;
   onOpenProjectSettings?: () => void;
 };
-
-type MemoryForm = {
-  title: string;
-  body: string;
-  trustLabel: string;
-  sourceKind: string;
-  sourceID: string;
-  enabled: boolean;
-};
-
-type SkillForm = {
-  title: string;
-  description: string;
-  trustLabel: string;
-};
-
-const MEMORY_TRUST_LABELS = [
-  "operator_memory",
-  "generated_summary",
-  "handoff",
-  "external_untrusted",
-  "runtime_state",
-];
-const MEMORY_SOURCE_KINDS = [
-  "operator",
-  "generated",
-  "generated_summary",
-  "task_output",
-  "chat_message",
-  "handoff",
-  "project_launch_context",
-  "external_handoff",
-  "runtime_state",
-];
 
 const PROJECTS_LIST_PANEL_WIDTH = 220;
 
@@ -3145,630 +3112,6 @@ function ProjectDecisionRow({
   );
 }
 
-function ProjectMemoryPanel({
-  candidates,
-  discoveringContext,
-  entries,
-  error,
-  loading,
-  onDiscoverContextSources,
-  onPromoteCandidate,
-  onRejectCandidate,
-  onDelete,
-  onEdit,
-  onNew,
-  onRefresh,
-  project,
-  rejectingCandidateID,
-}: {
-  candidates: ProjectMemoryCandidateRecord[];
-  discoveringContext: boolean;
-  entries: ProjectMemoryRecord[];
-  error: string;
-  loading: boolean;
-  onDiscoverContextSources: () => void;
-  onPromoteCandidate: (candidate: ProjectMemoryCandidateRecord) => void;
-  onRejectCandidate: (candidate: ProjectMemoryCandidateRecord) => void;
-  onDelete: (entry: ProjectMemoryRecord) => void;
-  onEdit: (entry: ProjectMemoryRecord) => void;
-  onNew: () => void;
-  onRefresh: () => void;
-  project: ProjectRecord | null;
-  rejectingCandidateID: string;
-}) {
-  if (!project) return null;
-  const enabledCount = entries.filter((entry) => entry.enabled).length;
-  const pendingCount = candidates.filter((candidate) => candidate.status === "pending").length;
-  const contextSources = project.context_sources ?? [];
-  return (
-    <div>
-      <div style={panelStyle}>
-        <div style={{ display: "flex", alignItems: "center", gap: 8, marginBottom: 10 }}>
-          <div>
-            <div style={sectionLabelStyle}>Memory / Context</div>
-            <div style={{ ...subtleTextStyle, marginTop: 3 }}>
-              {loading
-                ? "Loading project memory…"
-                : `${enabledCount} enabled / ${entries.length} saved entries · ${pendingCount} pending candidates`}
-            </div>
-          </div>
-          <button
-            className="btn btn-ghost btn-sm"
-            type="button"
-            aria-label="Refresh project memory"
-            title="Refresh"
-            onClick={onRefresh}
-            style={{ marginLeft: "auto" }}
-          >
-            <Icon d={Icons.refresh} size={12} />
-          </button>
-          <button
-            className="btn btn-ghost btn-sm"
-            type="button"
-            disabled={discoveringContext}
-            onClick={onDiscoverContextSources}
-          >
-            <Icon d={Icons.search} size={12} />
-            {discoveringContext ? "Discovering…" : "Discover"}
-          </button>
-          <button className="btn btn-primary btn-sm" type="button" onClick={onNew}>
-            <Icon d={Icons.plus} size={12} />
-            Memory
-          </button>
-        </div>
-        {error && (
-          <div style={{ marginBottom: 10 }}>
-            <InlineError message={error} />
-          </div>
-        )}
-        <div style={{ display: "grid", gap: 8, marginBottom: 12 }}>
-          <div style={sectionLabelStyle}>Workspace guidance</div>
-          {contextSources.length === 0 ? (
-            <div style={subtleTextStyle}>No context sources discovered or configured yet.</div>
-          ) : (
-            contextSources.map((source) => (
-              <ProjectContextSourceRow key={source.id} source={source} />
-            ))
-          )}
-        </div>
-        {candidates.length > 0 && (
-          <div style={{ display: "grid", gap: 8, marginBottom: 12 }}>
-            <div style={sectionLabelStyle}>Candidates</div>
-            {candidates.map((candidate) => (
-              <ProjectMemoryCandidateRow
-                key={candidate.id}
-                candidate={candidate}
-                pendingReject={rejectingCandidateID === candidate.id}
-                onPromote={() => onPromoteCandidate(candidate)}
-                onReject={() => onRejectCandidate(candidate)}
-              />
-            ))}
-          </div>
-        )}
-        {entries.length === 0 && !loading ? (
-          <div style={subtleTextStyle}>No project memory entries saved yet.</div>
-        ) : (
-          <div style={{ display: "grid", gap: 8 }}>
-            {entries.map((entry) => (
-              <ProjectMemoryRow
-                key={entry.id}
-                entry={entry}
-                onDelete={() => onDelete(entry)}
-                onEdit={() => onEdit(entry)}
-              />
-            ))}
-          </div>
-        )}
-      </div>
-    </div>
-  );
-}
-
-function ProjectSkillsPanel({
-  discovering,
-  error,
-  loading,
-  onDiscover,
-  onRefresh,
-  onUpdate,
-  project,
-  skills,
-  updatingSkillID,
-}: {
-  discovering: boolean;
-  error: string;
-  loading: boolean;
-  onDiscover: () => void;
-  onRefresh: () => void;
-  onUpdate: (skill: ProjectSkillRecord, patch: UpdateProjectSkillPayload) => void;
-  project: ProjectRecord | null;
-  skills: ProjectSkillRecord[];
-  updatingSkillID: string;
-}) {
-  if (!project) return null;
-  const enabledCount = skills.filter((skill) => skill.enabled).length;
-  const availableCount = skills.filter((skill) => skill.status === "available").length;
-  return (
-    <div>
-      <div style={panelStyle}>
-        <div style={{ display: "flex", alignItems: "center", gap: 8, marginBottom: 10 }}>
-          <div>
-            <div style={sectionLabelStyle}>Project Skills</div>
-            <div style={{ ...subtleTextStyle, marginTop: 3 }}>
-              {loading
-                ? "Loading project skills..."
-                : `${enabledCount} enabled / ${availableCount} available / ${skills.length} registered`}
-            </div>
-          </div>
-          <button
-            className="btn btn-ghost btn-sm"
-            type="button"
-            aria-label="Refresh project skills"
-            title="Refresh"
-            onClick={onRefresh}
-            style={{ marginLeft: "auto" }}
-          >
-            <Icon d={Icons.refresh} size={12} />
-          </button>
-          <button
-            className="btn btn-ghost btn-sm"
-            type="button"
-            disabled={discovering}
-            onClick={onDiscover}
-          >
-            <Icon d={Icons.search} size={12} />
-            {discovering ? "Discovering..." : "Discover"}
-          </button>
-        </div>
-        {error && (
-          <div style={{ marginBottom: 10 }}>
-            <InlineError message={error} />
-          </div>
-        )}
-        {skills.length === 0 && !loading ? (
-          <EmptyBlock
-            title="No project skills registered"
-            detail="Discover skills from AGENTS.md / CLAUDE.md references, .agents/skills, or .hecate/skills."
-          />
-        ) : (
-          <div style={{ display: "grid", gap: 8 }}>
-            {skills.map((skill) => (
-              <ProjectSkillRow
-                key={skill.id}
-                pending={updatingSkillID === skill.id}
-                skill={skill}
-                onUpdate={onUpdate}
-              />
-            ))}
-          </div>
-        )}
-      </div>
-    </div>
-  );
-}
-
-function ProjectSkillRow({
-  onUpdate,
-  pending,
-  skill,
-}: {
-  onUpdate: (skill: ProjectSkillRecord, patch: UpdateProjectSkillPayload) => void;
-  pending: boolean;
-  skill: ProjectSkillRecord;
-}) {
-  const [draft, setDraft] = useState(() => skillFormFromRecord(skill));
-
-  useEffect(() => {
-    setDraft(skillFormFromRecord(skill));
-  }, [skill]);
-
-  const changed =
-    draft.title.trim() !== skill.title ||
-    draft.description.trim() !== (skill.description ?? "") ||
-    draft.trustLabel.trim() !== skill.trust_label;
-  const statusClass = skill.status === "available" ? "badge badge-green" : "badge badge-amber";
-
-  return (
-    <div style={memoryEntryStyle}>
-      <div style={{ display: "flex", alignItems: "flex-start", gap: 10 }}>
-        <label
-          style={{
-            display: "inline-flex",
-            alignItems: "center",
-            gap: 6,
-            paddingTop: 2,
-            color: "var(--text1)",
-          }}
-        >
-          <input
-            type="checkbox"
-            checked={skill.enabled}
-            disabled={pending}
-            aria-label={`Enable skill ${skill.title || skill.id}`}
-            onChange={(event) => onUpdate(skill, { enabled: event.target.checked })}
-          />
-          <span className={skill.enabled ? "badge badge-green" : "badge badge-muted"}>
-            {skill.enabled ? "enabled" : "disabled"}
-          </span>
-        </label>
-        <div style={{ display: "grid", gap: 8, flex: 1, minWidth: 0 }}>
-          <div style={{ display: "flex", flexWrap: "wrap", gap: 6, alignItems: "center" }}>
-            <CopyableID text={skill.id} compact />
-            <span className={statusClass}>{skill.status}</span>
-            <span className="badge badge-muted">{skill.trust_label}</span>
-            <span className="badge badge-muted">{skill.format}</span>
-          </div>
-          <div style={{ display: "grid", gridTemplateColumns: "minmax(160px, 1fr) 1fr", gap: 8 }}>
-            <label style={fieldStyle}>
-              <span style={fieldLabelStyle}>Title</span>
-              <input
-                className="input"
-                value={draft.title}
-                disabled={pending}
-                onChange={(event) =>
-                  setDraft((current) => ({ ...current, title: event.target.value }))
-                }
-              />
-            </label>
-            <label style={fieldStyle}>
-              <span style={fieldLabelStyle}>Trust label</span>
-              <input
-                className="input"
-                value={draft.trustLabel}
-                disabled={pending}
-                onChange={(event) =>
-                  setDraft((current) => ({ ...current, trustLabel: event.target.value }))
-                }
-              />
-            </label>
-          </div>
-          <label style={fieldStyle}>
-            <span style={fieldLabelStyle}>Description</span>
-            <textarea
-              className="input"
-              rows={2}
-              value={draft.description}
-              disabled={pending}
-              onChange={(event) =>
-                setDraft((current) => ({ ...current, description: event.target.value }))
-              }
-            />
-          </label>
-          <div style={subtleTextStyle}>
-            {skill.path}
-            {skill.root_id ? ` · root ${skill.root_id}` : ""}
-            {skill.source_context_source_ids?.length
-              ? ` · sources ${skill.source_context_source_ids.join(", ")}`
-              : ""}
-            {skill.discovered_at ? ` · discovered ${formatAbsoluteTime(skill.discovered_at)}` : ""}
-          </div>
-          {skill.warnings?.length ? (
-            <div style={{ display: "grid", gap: 3 }}>
-              {skill.warnings.map((warning) => (
-                <div key={warning} style={{ ...subtleTextStyle, color: "var(--amber)" }}>
-                  {warning}
-                </div>
-              ))}
-            </div>
-          ) : null}
-        </div>
-        <button
-          className="btn btn-ghost btn-sm"
-          type="button"
-          disabled={pending || !changed}
-          onClick={() =>
-            onUpdate(skill, {
-              title: draft.title.trim(),
-              description: draft.description.trim(),
-              trust_label: draft.trustLabel.trim(),
-            })
-          }
-        >
-          {pending ? "Saving..." : "Save"}
-        </button>
-      </div>
-    </div>
-  );
-}
-
-function ProjectContextSourceRow({ source }: { source: ProjectContextSourceRecord }) {
-  const host = source.metadata?.host;
-  return (
-    <div style={memoryEntryStyle}>
-      <div style={{ display: "flex", alignItems: "center", gap: 8, minWidth: 0 }}>
-        <span
-          className={
-            source.kind === "workspace_instruction" ? "badge badge-green" : "badge badge-muted"
-          }
-        >
-          {source.kind}
-        </span>
-        <div style={{ ...titleStyle, flex: 1, minWidth: 0 }}>{source.title || source.path}</div>
-        <span className={source.enabled ? "badge badge-muted" : "badge badge-amber"}>
-          {source.enabled ? "enabled" : "disabled"}
-        </span>
-      </div>
-      <div style={metaLineStyle}>
-        <span>{source.path}</span>
-        {source.format && <span>{source.format}</span>}
-        {source.scope && <span>{source.scope}</span>}
-        {host && <span>{host}</span>}
-      </div>
-    </div>
-  );
-}
-
-function ProjectMemoryRow({
-  entry,
-  onDelete,
-  onEdit,
-}: {
-  entry: ProjectMemoryRecord;
-  onDelete: () => void;
-  onEdit: () => void;
-}) {
-  const source = formatMemorySource(entry);
-  return (
-    <div style={memoryEntryStyle}>
-      <div style={{ display: "flex", alignItems: "center", gap: 8, minWidth: 0 }}>
-        <span className={entry.enabled ? "badge badge-muted" : "badge badge-amber"}>
-          {entry.enabled ? "enabled" : "disabled"}
-        </span>
-        <span className="badge badge-muted">{entry.trust_label}</span>
-        <div style={{ ...titleStyle, flex: 1, minWidth: 0 }}>{entry.title}</div>
-        <button
-          className="btn btn-ghost btn-sm"
-          type="button"
-          aria-label={`Edit memory ${entry.title}`}
-          onClick={onEdit}
-          title="Edit"
-        >
-          <Icon d={Icons.edit} size={12} />
-        </button>
-        <button
-          className="btn btn-ghost btn-sm"
-          type="button"
-          aria-label={`Delete memory ${entry.title}`}
-          onClick={onDelete}
-          title="Delete"
-          style={{ color: "var(--red)" }}
-        >
-          <Icon d={Icons.trash} size={12} />
-        </button>
-      </div>
-      <div style={memoryBodyStyle}>{entry.body}</div>
-      <div style={metaLineStyle}>
-        <span>{source}</span>
-        <span>Updated {formatAbsoluteTime(entry.updated_at)}</span>
-        <CopyableID text={entry.id} compact />
-      </div>
-    </div>
-  );
-}
-
-function ProjectMemoryCandidateRow({
-  candidate,
-  onPromote,
-  onReject,
-  pendingReject,
-}: {
-  candidate: ProjectMemoryCandidateRecord;
-  onPromote: () => void;
-  onReject: () => void;
-  pendingReject: boolean;
-}) {
-  const source = formatCandidateSource(candidate);
-  const sourceRefs = formatCandidateSourceRefs(candidate);
-  const pending = candidate.status === "pending";
-  return (
-    <div style={memoryEntryStyle}>
-      <div style={{ display: "flex", alignItems: "center", gap: 8, minWidth: 0 }}>
-        <span className={pending ? "badge badge-amber" : "badge badge-muted"}>
-          {candidate.status}
-        </span>
-        <span className="badge badge-muted">{candidate.suggested_trust_label}</span>
-        <div style={{ ...titleStyle, flex: 1, minWidth: 0 }}>{candidate.title}</div>
-        {pending && (
-          <>
-            <button
-              className="btn btn-primary btn-sm"
-              type="button"
-              aria-label={`Promote memory candidate ${candidate.title}`}
-              onClick={onPromote}
-              title="Promote"
-            >
-              <Icon d={Icons.check} size={12} />
-            </button>
-            <button
-              className="btn btn-ghost btn-sm"
-              type="button"
-              aria-label={`Reject memory candidate ${candidate.title}`}
-              disabled={pendingReject}
-              onClick={onReject}
-              title="Reject"
-              style={{ color: "var(--red)" }}
-            >
-              <Icon d={Icons.x} size={12} />
-            </button>
-          </>
-        )}
-      </div>
-      <div style={memoryBodyStyle}>{candidate.body}</div>
-      <div style={metaLineStyle}>
-        <span>{source}</span>
-        <span>Suggested {formatAbsoluteTime(candidate.created_at)}</span>
-        <CopyableID text={candidate.id} compact />
-      </div>
-      {sourceRefs.length > 0 && (
-        <div style={{ ...subtleTextStyle, marginTop: 6 }}>
-          Source refs: {sourceRefs.join(" · ")}
-        </div>
-      )}
-    </div>
-  );
-}
-
-function ProjectMemoryModal({
-  candidate,
-  entry,
-  error,
-  pending,
-  onClose,
-  onSave,
-}: {
-  candidate?: ProjectMemoryCandidateRecord | null;
-  entry: ProjectMemoryRecord | null;
-  error: string;
-  pending: boolean;
-  onClose: () => void;
-  onSave: (form: MemoryForm) => void | Promise<void>;
-}) {
-  const [form, setForm] = useState<MemoryForm>(() =>
-    candidate ? memoryFormFromCandidate(candidate) : memoryFormFromRecord(entry),
-  );
-  const valid = form.title.trim().length > 0 && form.body.trim().length > 0;
-  const isCandidate = Boolean(candidate);
-  const candidateSourceRefs = candidate ? formatCandidateSourceRefs(candidate) : [];
-  return (
-    <Modal
-      title={
-        isCandidate
-          ? "Promote memory candidate"
-          : entry
-            ? "Edit project memory"
-            : "New project memory"
-      }
-      onClose={onClose}
-      width={620}
-      footer={
-        <button
-          className="btn btn-primary"
-          type="button"
-          disabled={pending || !valid}
-          onClick={() => void onSave(form)}
-          style={{ width: "100%", justifyContent: "center" }}
-        >
-          {pending
-            ? "Saving…"
-            : isCandidate
-              ? "Promote memory"
-              : entry
-                ? "Save memory"
-                : "Create memory"}
-        </button>
-      }
-    >
-      <form
-        onSubmit={(event) => {
-          event.preventDefault();
-          if (valid) void onSave(form);
-        }}
-        style={{ display: "grid", gap: 12 }}
-      >
-        {error && <InlineError message={error} />}
-        {candidate && (
-          <div
-            style={{
-              background: "var(--bg2)",
-              border: "1px solid var(--border)",
-              borderRadius: "var(--radius-sm)",
-              display: "grid",
-              gap: 6,
-              padding: "9px 10px",
-            }}
-          >
-            <div style={sectionLabelStyle}>Candidate provenance</div>
-            <div style={metaLineStyle}>
-              <span>{formatCandidateSource(candidate)}</span>
-              <span>{candidate.suggested_trust_label}</span>
-              <span>{candidate.status}</span>
-            </div>
-            {candidateSourceRefs.length > 0 && (
-              <div style={{ ...subtleTextStyle }}>
-                Source refs: {candidateSourceRefs.join(" · ")}
-              </div>
-            )}
-          </div>
-        )}
-        <label style={fieldStyle}>
-          <span style={fieldLabelStyle}>Title</span>
-          <input
-            className="input"
-            autoFocus
-            value={form.title}
-            onChange={(event) => setForm((current) => ({ ...current, title: event.target.value }))}
-          />
-        </label>
-        <label style={fieldStyle}>
-          <span style={fieldLabelStyle}>Body</span>
-          <textarea
-            className="input"
-            value={form.body}
-            rows={7}
-            onChange={(event) => setForm((current) => ({ ...current, body: event.target.value }))}
-            style={{ resize: "vertical", minHeight: 150 }}
-          />
-        </label>
-        <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr", gap: 10 }}>
-          <label style={fieldStyle}>
-            <span style={fieldLabelStyle}>Trust label</span>
-            <select
-              className="input"
-              value={form.trustLabel}
-              onChange={(event) =>
-                setForm((current) => ({ ...current, trustLabel: event.target.value }))
-              }
-            >
-              {MEMORY_TRUST_LABELS.map((label) => (
-                <option key={label} value={label}>
-                  {label}
-                </option>
-              ))}
-            </select>
-          </label>
-          <label style={fieldStyle}>
-            <span style={fieldLabelStyle}>Source kind</span>
-            <select
-              className="input"
-              value={form.sourceKind}
-              onChange={(event) =>
-                setForm((current) => ({ ...current, sourceKind: event.target.value }))
-              }
-            >
-              {MEMORY_SOURCE_KINDS.map((kind) => (
-                <option key={kind} value={kind}>
-                  {kind}
-                </option>
-              ))}
-            </select>
-          </label>
-        </div>
-        <label style={fieldStyle}>
-          <span style={fieldLabelStyle}>Source ID</span>
-          <input
-            className="input"
-            value={form.sourceID}
-            onChange={(event) =>
-              setForm((current) => ({ ...current, sourceID: event.target.value }))
-            }
-            placeholder="optional artifact, chat, message, or handoff id"
-          />
-        </label>
-        <label style={{ display: "flex", alignItems: "center", gap: 8, color: "var(--t1)" }}>
-          <input
-            type="checkbox"
-            checked={form.enabled}
-            onChange={(event) =>
-              setForm((current) => ({ ...current, enabled: event.target.checked }))
-            }
-          />
-          Enabled for project context packets
-        </label>
-      </form>
-    </Modal>
-  );
-}
-
 function WorkItemDetail({
   activityByAssignmentID,
   assignments,
@@ -4778,14 +4121,6 @@ function summarizeAssignments(assignments: ProjectAssignmentRecord[]): WorkItemS
   );
 }
 
-function skillFormFromRecord(skill: ProjectSkillRecord): SkillForm {
-  return {
-    title: skill.title ?? "",
-    description: skill.description ?? "",
-    trustLabel: skill.trust_label ?? "workspace_skill",
-  };
-}
-
 function linkedChatStatusTitle(linkedChat: NonNullable<ProjectActivityItemRecord["linked_chat"]>) {
   return [
     linkedChat.title,
@@ -4813,57 +4148,6 @@ function handoffSourceRefs(handoff: ProjectHandoffRecord): string[] {
     }
   }
   return refs;
-}
-
-function memoryFormFromRecord(entry: ProjectMemoryRecord | null): MemoryForm {
-  return {
-    title: entry?.title ?? "",
-    body: entry?.body ?? "",
-    trustLabel: entry?.trust_label ?? "operator_memory",
-    sourceKind: entry?.source_kind ?? "operator",
-    sourceID: entry?.source_id ?? "",
-    enabled: entry?.enabled ?? true,
-  };
-}
-
-function memoryFormFromCandidate(candidate: ProjectMemoryCandidateRecord): MemoryForm {
-  return {
-    title: candidate.title,
-    body: candidate.body,
-    trustLabel: candidate.suggested_trust_label || "generated_summary",
-    sourceKind: candidate.suggested_source_kind || "generated",
-    sourceID: candidate.suggested_source_id ?? "",
-    enabled: true,
-  };
-}
-
-function formatMemorySource(entry: ProjectMemoryRecord): string {
-  const sourceKind = entry.source_kind || "operator";
-  return entry.source_id ? `${sourceKind}:${entry.source_id}` : sourceKind;
-}
-
-function formatCandidateSource(candidate: ProjectMemoryCandidateRecord): string {
-  const refs = candidate.source_refs ?? [];
-  if (refs.length > 0) {
-    const ref = refs[0];
-    const label = ref.title || ref.id || ref.kind;
-    const suffix = refs.length > 1 ? ` +${refs.length - 1}` : "";
-    return `${ref.kind}:${label}${suffix}`;
-  }
-  const sourceKind = candidate.suggested_source_kind || "generated";
-  return candidate.suggested_source_id
-    ? `${sourceKind}:${candidate.suggested_source_id}`
-    : sourceKind;
-}
-
-function formatCandidateSourceRefs(candidate: ProjectMemoryCandidateRecord): string[] {
-  return (candidate.source_refs ?? [])
-    .map((ref) => {
-      const label = ref.title || (ref.id ? shortID(ref.id) : ref.url || "");
-      if (!label) return ref.kind;
-      return `${ref.kind} ${label}`;
-    })
-    .filter(Boolean);
 }
 
 function buildProjectAssignmentChatLaunchRequest({
@@ -5282,18 +4566,6 @@ const metaLineStyle: CSSProperties = {
   color: "var(--t3)",
   fontSize: 11,
   marginTop: 6,
-};
-
-const fieldStyle: CSSProperties = {
-  display: "grid",
-  gap: 6,
-};
-
-const fieldLabelStyle: CSSProperties = {
-  color: "var(--t2)",
-  fontFamily: "var(--font-mono)",
-  fontSize: 11,
-  textTransform: "uppercase",
 };
 
 const panelStyle: CSSProperties = {
@@ -5806,18 +5078,4 @@ const activityBucketButtonStyle: CSSProperties = {
 const artifactStyle: CSSProperties = {
   borderTop: "1px solid var(--border)",
   paddingTop: 8,
-};
-
-const memoryEntryStyle: CSSProperties = {
-  borderTop: "1px solid var(--border)",
-  paddingTop: 8,
-};
-
-const memoryBodyStyle: CSSProperties = {
-  marginTop: 6,
-  color: "var(--t1)",
-  fontSize: 12,
-  lineHeight: 1.45,
-  whiteSpace: "pre-wrap",
-  overflowWrap: "anywhere",
 };
